Add configuration settings/preferences
This adds administrator-level configuration settings, as well as
user-level preferences, to control the number of courses displayed,
the order in which courses are displayed, and how grades are
formatted/displayed. All the settings can be locked at the
administrator level. The order in which courses is displayed is
either descending by course id or using the user-set sort order
from the Course Overview block. Grades can now be formatted nine
different ways.

This also brings variable names and such up to current Moodle
standards.